refactor: address OCR findings; record the 89.7% coverage baseline
OCR (scoped to the vendor/telemetry diff, 16 files) found two low items:
- AidaFrameSnapshot was a record over a byte[] payload, implying a value
  equality that is false here (the bytes alias the reader's reusable buffer).
  It is now a sealed class with get-only properties and the lifetime documented.
- FakeAidaMmapSource compared offset + length in 32-bit, so a producer-controlled
  offset near int.MaxValue could wrap and throw out of Array.Copy, diverging from
  the production 64-bit guard. Fixed.

Coverage rerun (measure-coverage.ps1): 89.7% of the 16,834 instrumented lines
(2,465 tests), was 90.7% at 2,407 tests; the move is the new vendor/AIDA/
telemetry surface. The gate floors (Sdk/Core/Hardware >= 70%) all hold with
margin, and CONTEXT now records the 2026-09-14 numbers.

+++ ModernWigiDash.Tests/TestDoubles.cs
@@ -305,7 +305,10 @@ internal sealed class FakeAidaMmapSource(byte[]? map = null) : IAidaMmapSource
             return false;
         }
 
-        if (offset < 0 || length < 0 || offset + length > Map.Length || length > destination.Length)
+        // 64-bit compare, matching the production adapter (a producer-controlled
+        // offset near int.MaxValue wrapped the 32-bit sum negative and reached
+        // Array.Copy, which throws — the seam must return false instead).
+        if (offset < 0 || length < 0 || (long)offset + length > Map.Length || length > destination.Length)
         {
             error = "range out of bounds";
             return false;
